feat(frontend): convert add-model modal flows to multi-step drawer
Replace ProviderSelectionDialog + ConnectionDialog (both MUI Dialogs)
with a single multi-step BaseDrawer (ModelConnectionDrawer):

- Step 1 (select): provider list panel, cancel only
- Step 2 (configure): connection form with Back button in create flow
- Edit flow (model card click): opens directly at configure step

New components:
- ProviderSelectionPanel: content-only provider list
- ConnectionForm: extracted form body, exposes submit() via ref,
  lifts loading/canSave state to parent drawer
- ModelConnectionDrawer: multi-step BaseDrawer orchestrating both steps

Remove old ConnectionDialog, models/ProviderSelectionDialog, and
common/ProviderSelectionDialog.
