+++ apps/frontend/src/app/(protected)/models/components/index.ts
@@ -1,3 +1,4 @@
-export { ProviderSelectionDialog } from './ProviderSelectionDialog';
-export { ConnectionDialog } from './ConnectionDialog';
 export { ConnectedModelCard } from './ModelCard';
+export { ModelConnectionDrawer } from './ModelConnectionDrawer';
+export { ProviderSelectionPanel } from './ProviderSelectionPanel';
+export { ConnectionForm } from './ConnectionForm';
